requirements.txt
fastapi
uvicorn
python-dotenv
groq
qdrant-client
pydantic
sentence-transformers
pydantic[email]